+++ src/components/Upload/src/props.ts
@@ -14,11 +14,13 @@ type SortableOptions = Merge<
     // ...可扩展
   }
 >;
-
+type previewColumnsFnType = {
+  handleRemove:(record:Record<string,any>,key:string)=>any,
+  handleAdd:(record:Record<string,any>,key:string)=>any,
+}
 export const previewType = {
   previewColumns: {
-    type: Array as PropType<BasicColumn[] | FileBasicColumn[]>,
-    default: [],
+    type: [Array,Function] as PropType<BasicColumn[] | ((arg:previewColumnsFnType) => BasicColumn[])>,
     required: false,
   },
   beforePreviewData: {
@@ -110,6 +112,10 @@ export const previewProps = {
   value: {
     type: Array as PropType<string[]>,
     default: () => [],
+  },
+  maxNumber: {
+    type: Number as PropType<number>,
+    default: 1,
   },
   ...previewType,
 };
